Commented the "rvest" method out.
Commented the "rvest" method out as the loop does not work as intended, and breaks somewhere around the 45-50th iteration. If you can get it to work, feel free to use it—I've left it in the document for reference, but just be aware that as it stands, it does not function as intended.